+++ src/Export/ExporterInterface.php
@@ -14,7 +14,6 @@ interface ExporterInterface
      * @param StreamInterface       $stream
      * @param DataProviderInterface $dataProvider
      * @param array                 $columns
-     * @param CollectionInterface   $collection
      */
-    public function exportEntities(StreamInterface $stream, DataProviderInterface $dataProvider, array $columns, CollectionInterface $collection): void;
+    public function exportEntities(StreamInterface $stream, DataProviderInterface $dataProvider, array $columns): void;
 }